Revert "Merge react/ui branch work into develop "
