+++ react/01_m-basic/public/manifest.json
@@ -1,9 +1,9 @@
 {
-  "short_name": "m-evergreen",
-  "name": "m-evergreen",
+  "short_name": "React App",
+  "name": "Create React App Sample",
   "icons": [
     {
-      "src": "favicon16.ico",
+      "src": "favicon.ico",
       "sizes": "64x64 32x32 24x24 16x16",
       "type": "image/x-icon"
     }
