+++ react/01_m-basic/src/App.css
@@ -1,0 +1,9 @@
+
+/*
+  src/App.css
+  app component style sheet
+*/
+
+.App {
+  text-align: center;
+}
