+++ react/01_m-basic/src/App.js
@@ -1,0 +1,22 @@
+
+/*
+  src/App.js
+*/
+
+// import react component library
+import React, { Component } from 'react';
+// import app component styles
+import './App.css';
+
+// App component class
+class App extends Component {
+  render() {
+    return (
+      <div className="App">
+        <h1> App Component </h1>
+      </div>
+    );
+  }
+}
+
+export default App;
