--- react/01_m-basic/src/app/components/Jumbo/Jumbo.css
@@ -1,23 +1,0 @@
-
-/*
-  src/app/components/Jumbo/Jumbo.css
-  jumbotron component styles
-*/
-
-.Jumbo {
-  text-align: center;
-  margin: 0 auto;
-}
-
-#tron {
-  background: #D5D1D6;
-}
-
-/* Extra Small Devices, Phones */
-@media(min-width:480px){
-	#tron > Image {
-		margin-top: -2em;
-		width: 99vw;
-	  /*display: none;*/
-	}
-}
