+++ react/01_m-basic/src/components/Jumbo/Jumbo.css
@@ -1,0 +1,21 @@
+
+/* Jumbo react component style sheet */
+
+.Jumbo {
+  text-align: center;
+  margin: 0 auto;
+  padding-top: .5em;
+}
+
+#tron {
+  background: #D5D1D6;
+}
+
+/* Extra Small Devices, Phones */
+@media(min-width:480px){
+	#tron > Image {
+		margin-top: -2em;
+		width: 99vw;
+	  /*display: none;*/
+	}
+}
