+++ react/01_m-basic/src/components/Jumbo/Jumbo.test.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import ReactDOM from 'react-dom';
+import Jumbo from './Jumbo';
+
+it('renders without crashing', () => {
+  const div = document.createElement('div');
+  ReactDOM.render(<Jumbo />, div);
+  ReactDOM.unmountComponentAtNode(div);
+});
