--- react/02_m-basic-ui/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "m-basic-ui",
-  "version": "0.2",
-  "private": true,
-  "dependencies": {
-    "evergreen-ui": "^4.2.0",
-    "react": "^16.6.0",
-    "react-bootstrap": "^0.32.4",
-    "react-dom": "^16.6.0",
-    "react-scripts": "2.1.0"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
-}
