--- react/02_m-basic-ui/public/index.html
@@ -1,38 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head> <!-- START of HEAD -->
-    <!-- PROPERTIES -->
-    <meta charset="utf-8">
-    <!-- favicon -->
-    <link rel="shortcut icon" href="%PUBLIC_URL%/favicon16.ico">
-    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
-    <!-- FOR MOBILE
-      <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
-    -->
-    <meta name="theme-color" content="#000000">
-    <!--
-      manifest.json provides metadata used when your web app is added to the
-      homescreen on Android. See https://developers.google.com/web/fundamentals/web-app-manifest/
-    -->
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json">
-    <!-- Application Title -->
-    <title>M-Basic-UI</title>
-    <!-- Bootstrap 3 - compiled and minified CSS -->
-    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"
-          integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
-  </head> <!-- END of HEAD -->
-  <body> <!-- START of BODY -->
-    <noscript>
-      You need to enable JavaScript to run this app.
-    </noscript>
-    <!-- Jumbotron header component -->
-    <div id="jumbo"></div>
-    <!-- root app component -->
-    <div id="app-root"></div>
-    <!-- UI component -->
-    <div id="ui"></div>
-  </body> <!-- END of BODY -->
-</html>
-
-
-<!-- EOF -->
