--- react/02_m-basic-ui/src/app/App/App.js
@@ -1,23 +1,0 @@
-
-/*
-  src/app/App/App.js
-  app react component
-*/
-
-// import react component library
-import React, { Component } from 'react';
-// import app component styles
-import './App.css';
-
-// App component class
-class App extends Component {
-  render() {
-    return (
-      <div className="App">
-        <h2> -- 02 Basic UI -- </h2>
-      </div>
-    );
-  }
-}
-
-export default App;
