--- react/02_m-basic-ui/src/index.css
@@ -1,24 +1,0 @@
-
-/*
-  src/index.css
-  main application style sheet
-*/
-
-body {
-  margin: 0;
-  padding: 0;
-  background: #D5D1D6;
-  font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", "Roboto", "Oxygen",
-    "Ubuntu", "Cantarell", "Fira Sans", "Droid Sans", "Helvetica Neue", sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  /*
-  min-width: 500px;
-  width: auto !important; /* Firefox will set width as auto *
-  width: 500px;
-  */
-}
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New", monospace;
-}
