+++ react/02_m-evergreen/public/index.html
@@ -16,7 +16,7 @@
     -->
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json">
     <!-- Application Title -->
-    <title>M8Fr8</title>
+    <title>m-evergreen</title>
     <!-- Bootstrap 3 - compiled and minified CSS -->
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css"
           integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous">
@@ -25,10 +25,12 @@
     <noscript>
       You need to enable JavaScript to run this app.
     </noscript>
-    <!-- Jumbotron component -->
+    <!-- Jumbotron header component -->
     <div id="jumbo"></div>
     <!-- root app component -->
     <div id="root"></div>
+    <!-- evergreen ui component -->
+    <div id="evergreen"></div>
   </body> <!-- END of BODY -->
 </html>
