+++ react/02_m-evergreen/public/manifest.json
