+++ react/02_m-evergreen/src/app/app/App.css
@@ -1,0 +1,13 @@
+
+/*
+  src/app/App/App.css
+  app component style sheet
+*/
+
+.App {
+  text-align: center;
+  margin: .5em auto;
+  padding-bottom: 1em;
+  max-width: 550px;
+  border: 1px dotted red;
+}
