+++ react/02_m-evergreen/src/index.js
@@ -11,9 +11,11 @@ import ReactDOM from 'react-dom';
 import './index.css';
 
 // import jumbo component
-import Jumbo from './components/Jumbo/Jumbo';
+import Jumbo from './app/components/Jumbo/Jumbo';
 // import app component
-import App from './App';
+import App from './app/App/App';
+// import evergreen component
+import Evergreen from './app/components/Evergreen/Evergreen';
 
 // import service worker
 import * as serviceWorker from './serviceWorker';
@@ -22,6 +24,8 @@ import * as serviceWorker from './serviceWorker';
 ReactDOM.render(<Jumbo />, document.getElementById('jumbo'));
 // render app component
 ReactDOM.render(<App />, document.getElementById('root'));
+// render evergreen component
+ReactDOM.render(<Evergreen />, document.getElementById('evergreen'));
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
