--- react/03_m_ui/public/manifest.json
@@ -1,15 +1,0 @@
-{
-  "short_name": "m-ui",
-  "name": "m-ui",
-  "icons": [
-    {
-      "src": "favicon16.ico",
-      "sizes": "64x64 32x32 24x24 16x16",
-      "type": "image/x-icon"
-    }
-  ],
-  "start_url": ".",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
