--- react/03_m_ui/src/app/components/Jumbo/Jumbo.css
@@ -1,25 +1,0 @@
-
-/*
-  src/app/components/Jumbo/Jumbo.css
-  jumbotron component styles
-*/
-
-.Jumbo {
-  text-align: center;
-  margin: 0 auto;
-  /*max-width: 500px;*/
-  margin-top: -1em;
-}
-
-#tron {
-  background: #D5D1D6;
-}
-
-/* Extra Small Devices, Phones */
-/*display: none;*/
-@media(min-width:480px){
-	#tron > Image {
-		margin-top: -2em;
-		width: 99vw;
-	}
-}
